+++ images/pale_blue/icon_gba.h
@@ -3,11 +3,11 @@
 
 //======================================================================
 //
-//	gImage_icon_gba_PALE_BLUE, 16x14@16,
+//	gImage_icon_gba_PALE_BLUE, 16x14@16, 
 //	+ bitmap not compressed
 //	Total size: 448 = 448
 //
-//	Time-stamp: 2026-06-08, 09:52:04
+//	Time-stamp: 2026-07-04, 11:44:18
 //	Exported by Cearn's GBA Image Transmogrifier, v0.8.6
 //	( http://www.coranac.com/projects/#grit )
 //